feat(gpu): add merkle kernels to bb31 hasher
